add audio cues and alert with response
