+++ src/vs/workbench/contrib/chat/browser/chatWidget.ts
@@ -4,7 +4,6 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as dom from 'vs/base/browser/dom';
-import { status } from 'vs/base/browser/ui/aria/aria';
 import { ITreeContextMenuEvent, ITreeElement } from 'vs/base/browser/ui/tree/tree';
 import { CancellationToken } from 'vs/base/common/cancellation';
 import { Emitter } from 'vs/base/common/event';
@@ -16,7 +15,6 @@ import 'vs/css!./media/chat';
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
 import { localize } from 'vs/nls';
 import { MenuId } from 'vs/platform/actions/common/actions';
-import { AudioCue, AudioCueGroupId, IAudioCueService } from 'vs/platform/audioCues/browser/audioCueService';
 import { IContextKey, IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
 import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
@@ -518,30 +516,3 @@ export class ChatWidgetService implements IChatWidgetService {
 		);
 	}
 }
-
-
-const CHAT_RESPONSE_PENDING_AUDIO_CUE_LOOP_MS = 5000;
-export class ChatAccessibilityService extends Disposable implements IChatAccessibilityService {
-
-	declare readonly _serviceBrand: undefined;
-
-	private _responsePendingAudioCue: IDisposable | undefined;
-
-	constructor(@IAudioCueService private readonly _audioCueService: IAudioCueService) {
-		super();
-	}
-	acceptRequest(): void {
-		this._audioCueService.playAudioCue(AudioCue.chatRequestSent, true);
-		this._responsePendingAudioCue = this._audioCueService.playAudioCueLoop(AudioCue.chatResponsePending, CHAT_RESPONSE_PENDING_AUDIO_CUE_LOOP_MS);
-	}
-	acceptResponse(response?: IChatResponseViewModel): void {
-		this._responsePendingAudioCue?.dispose();
-		this._audioCueService.playRandomAudioCue(AudioCueGroupId.chatResponseReceived, true);
-		if (!response) {
-			return;
-		}
-		const errorDetails = response.errorDetails ? ` ${response.errorDetails.message}` : '';
-		status(response.response.value + errorDetails);
-	}
-}
-
